pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vaibvoice"
version = "1.0.0"
description = "VaibVoice - Voice transcription application"
readme = "README.md"
authors = [
    {name = "VaibVoice Team", email = "info@vaibvoice.com"}
]
license = {text = "MIT"}
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
requires-python = ">=3.8"
dependencies = [
    "sounddevice",
    "soundfile",
    "numpy",
    "requests",
    "rich",
    "python-dotenv",
    "pynput",
    "pyautogui",
    "openai",
    "fastapi",
    "uvicorn",
    "pydantic",
]

[project.scripts]
vaibvoice = "vaibvoice.main:main"

[tool.setuptools]
packages = ["vaibvoice"]

[tool.black]
line-length = 88
target-version = ["py38"]

[tool.isort]
profile = "black"
line_length = 88